User, Story, Vote models + basic front page
Note that the front page will appear blank until you create:
- a user
- a story associated with that user

and that the vote count will not be displayed untless you manually set `votes`
on the Story instance.

A basic `time_formatter` was also added to the application helper for a somewhat
friendly version of time on the home index.